Add ui.js file, modularize code
